% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compareWithTG.R
\name{compareWithTG}
\alias{compareWithTG}
\title{Extract MDT values at TG positions}
\usage{
compareWithTG(TG, dat, lonlim, latlim, boxlon = 3, boxlat = 3)
}
\arguments{
\item{TG}{...}

\item{dat}{...}

\item{lonlim}{...}

\item{latlim}{...}

\item{boxlon}{}

\item{boxlat}{}
}
\value{
list with mean and sd values
##'
}
\description{
Extract MDT values at TG positions
}
\details{
...
}

