\name{dDTU15MSS_ref2003-2007}
\alias{dDTU15MSS_ref2003-2007}
\docType{data}
\title{
dDTU15MSS_ref2003-2007
}
\description{
to be written
}
\usage{data("dDTU15MSS_ref2003-2007")}
\format{
  A data frame with 8476 observations on the following 4 variables.
  \describe{
    \item{\code{height}}{a numeric vector}
    \item{\code{lake}}{a numeric vector}
    \item{\code{track}}{a numeric vector}
    \item{\code{time}}{a numeric vector}
  }
}
\details{
...
}
\source{
...
}
\references{
...
}
\examples{
data(dDTU15MSS_ref2003-2007)
}
\keyword{datasets}
