\name{difmss15eig6c4r}
\alias{difmss15eig6c4r}
\docType{data}
\title{
difmss15eig6c4r
}
\description{
difmss15eig6c4r is an unfiltered MDT based on the MSS DTU15 and the
geoid model Eigen6c4r. difmss15eig6c4r is a 2880 x 1440 matrix. The first cell is longitude 0 to 1/8 degree, latitude -90 to -90+1/8 degree
cells are all 1/8 by 1/8 degree
order is east to west, then south to north. The MDT values a in meters
and are referenced to the TOPEX ellipsoid. The grid is in the permanent
mean tide system. 
}
\usage{data("difmss15eig6c4r")}
\format{
  A matrix[lon,lat] of dimension 2880 x 1440
}
\details{
...
}
\source{
...
}
\references{
Created by Per Knudsen
}
\examples{
data(difmss15eig6c4r)
}
\keyword{datasets}
