\name{DTU15MSS}
\alias{DTU15MSS}
\docType{data}
\title{
DTU15MSS
}
\description{
to be written
}
\usage{data("DTU15MSS")}
\format{
  A data frame with 8476 observations on the following 4 variables.
  \describe{
    \item{\code{height}}{a numeric vector}
    \item{\code{lake}}{a numeric vector}
    \item{\code{track}}{a numeric vector}
    \item{\code{time}}{a numeric vector}
  }
}
\details{
...
}
\source{
...
}
\references{
...
}
\examples{
data(DTU15MSS)
}
\keyword{datasets}
