\name{eigen6c4r}
\alias{eigen6c4r}
\docType{data}
\title{
eigen6c4r
}
\description{
to be written
}
\usage{data("eigen6c4r")}
\format{
  A data frame with 8476 observations on the following 4 variables.
  \describe{
    \item{\code{height}}{a numeric vector}
    \item{\code{lake}}{a numeric vector}
    \item{\code{track}}{a numeric vector}
    \item{\code{time}}{a numeric vector}
  }
}
\details{
...
}
\source{
...
}
\references{
...
}
\examples{
data(eigen6c4r)
}
\keyword{datasets}
