% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ellipsoidTF2MT.R
\name{ellipsoidTF2MT}
\alias{ellipsoidTF2MT}
\title{Function that estimates the correction to go from a tide free ellipsoid to a mean tide ellipsoid.}
\usage{
ellipsoidTF2MT(phi, h2 = 0.62)
}
\arguments{
\item{phi}{The latitude in degrees.}

\item{h2}{Love number, the default value is h2=0.62}
}
\value{
An array with the height differences in meter.
}
\description{
Function that estimates the correction to go from a tide free ellipsoid to a mean tide ellipsoid.
}
\details{
...
}

