% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gaussAniKernel.R
\name{gaussAniKernel}
\alias{gaussAniKernel}
\title{Anisotropic gauss kernel}
\usage{
gaussAniKernel(sigmau, sigmav, ny, nx)
}
\arguments{
\item{sigmau}{...}

\item{sigmav}{...}

\item{ny}{...}

\item{nx}{...}
}
\description{
Anisotropic gauss kernel
}
\details{
...
}

