% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helper.R
\name{getBoxMean}
\alias{getBoxMean}
\title{Find mean value in a box}
\usage{
getBoxMean(dat, id, boxlon = 4, boxlat = 4)
}
\arguments{
\item{dat}{...}

\item{id}{...}

\item{boxlon}{...}

\item{boxlat}{...}
}
\value{
list with mean and sd values
##'
}
\description{
Find mean value in a box
}
\details{
...
}

