% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getCoastLine.R
\name{getCoastLine}
\alias{getCoastLine}
\title{Extract sub grid}
\usage{
getCoastLine(mask, land = 0, water = 1)
}
\arguments{
\item{mask}{Matrix[lon,lat] with land mask values.}

\item{land}{Integer value for land. Default is 0.}

\item{water}{Integer value for land. Default is 0.}
}
\value{
List with the elements;  matrix g[lon,lat], which contains the location of the land value which has a water neighbor. Matrix id which contains two columns; row no and col no of the matrix g where a land value is identified as coast.
}
\description{
Extract sub grid
}
\details{
...
}

