% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCount.R
\name{getCount}
\alias{getCount}
\title{Helper function for the bootstrap function getError}
\usage{
getCount(IMat, nr, nc)
}
\arguments{
\item{IMat}{a matrix that divides the area into sub areas. It is the output of the function getIdMat.}

\item{nr}{the number of rows in the data matrix (the raw MDT).}

\item{nc}{the number of columns in the data matrix (the raw MDT)}
}
\value{
matrix[lon,lat] where each element represents the number of times each elements should be counted in the iterative filter, when the sd is estimated for the filtered MDT, by the use of bootstrap.
}
\description{
This function returns a matrix with the number of times each element must be counted.
}
\details{
...
}
