% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getError.R
\name{getError}
\alias{getError}
\title{Estimate a MDT error field via bootstrap}
\usage{
getError(dat, land, mask, bootNr = 100, nnx = 3, nny = 3,
  ncores = detectCores())
}
\arguments{
\item{dat}{An object as returned by the function 'getSubGrid' or 'iterativeAveSmoother', which includes a list containing  a matrix g[lon,lat], a vector lon (longitudes) and a vector lat (latitudes).}

\item{land}{Matrix[lon,lat] containing land values}

\item{mask}{An object as returned by the function 'getSubGrid', which includes a list containing  a matrix g[lon,lat], a vector lon (longitudes) and a vector lat (latitudes). mask$g is a Matrix[lon,lat] representing the land mask, where land=0 and water=1.}

\item{bootNr}{Number of bootstap data sets}

\item{nnx}{number of grid cells in the east-west direction}

\item{nny}{number of grid cells in the north-south direction}

\item{ncores}{Number of available cores}

\item{dat}{An object as returned by the function 'getSubGrid', which includes a list containing  a matrix g[lon,lat], a vector lon (longitudes) and a vector lat (latitudes). The matrix dat$g[lon,lat] containes the values to be filtered.}
}
\description{
With the function getError it is possible to estimate an MDT error field via the bootstrap approach.
}
\details{
The data is sampled in blocks, which size is specified by the user in the arguments nnx and nny. nnx and nny should be chosen carefully. If nnx and nny are too small the data will not be independent and the error will be underestimated. If too large, there will be data gabs and the errors will not be representative.
}
