% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIdMat.R
\name{getIdMat}
\alias{getIdMat}
\title{Helper function for getError}
\usage{
getIdMat(mask, nnx, nny, nr, nc)
}
\arguments{
\item{mask}{An object as returned by the function 'getSubGrid', which includes a list containing  a matrix g[lon,lat], a vector lon (longitudes) and a vector lat (latitudes). mask$g is a Matrix[lon,lat] representing the land mask, where land=0 and water=1.}

\item{nnx}{number of grid cells in the east-west direction}

\item{nny}{number of grid cells in the north-south direction}

\item{nr}{number of rows in the data matrix}

\item{nc}{number of columns in the data matrix}
}
\value{
List with the elements;  matrix[lon,lat] Mat (ids of the submatrices), vector mysamples (the values of the ids), nrSam (the length of mysamples) and idland, the land index.
}
\description{
The function getIdMat divides a matrix into sub matrices specified by the arguments  nnx and nny .
}
\details{

}
