% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getLandComb.R
\name{getLandComb}
\alias{getLandComb}
\title{Extract MDT values at TG positions}
\usage{
getLandComb(polyCoast, TG, TGcorr, dat, lonlim, latlim, boxlon = 4,
  boxlat = 4)
}
\arguments{
\item{polyCoast}{...}

\item{TG}{...}

\item{TGcorr}{...}

\item{dat}{...}

\item{lonlim}{...}

\item{latlim}{...}

\item{boxlon}{...}

\item{boxlat}{...}
}
\value{
...
##'
}
\description{
Extract MDT values at TG positions
}
\details{
...
}

