% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getLandInfo.R
\name{getLandInfo}
\alias{getLandInfo}
\title{Helper function to getLandVal: Estimate land MDT values at the coastline based on altimetry}
\usage{
getLandInfo(mycoast, mask, dat, boxlon = 4, boxlat = 4)
}
\arguments{
\item{mycoast}{Matrix[N,2] with row and column values of the coast line. mycoast is the out put of helper function getCoastLine}

\item{mask}{Matrix[lon,lat] representing the land mask, where land=0 and water=1}

\item{dat}{Matrix[lon,lat] with MDT values}

\item{boxlon}{Integer. The number ((2 x boxlon) +1) of grid cells in the longitude direction, that is used to estimate the altimetry based MDT value at the coast.}

\item{boxlat}{Integer. The number ((2 x boxlat) +1) of grid cells in the latitude direction, that is used to estimate the altimetry based MDT value at the coast.}
}
\value{
Matrix[N,4]; row id, col id, mean MDT value, sd of MDT
}
\description{
Helper function to getLandVal: Estimate land MDT values at the coastline based on altimetry
}
\details{
...
}

