% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getLandVal.R
\name{getLandVal}
\alias{getLandVal}
\title{Extract MDT values at TG positions}
\usage{
getLandVal(dat, TG, mask, lonlim, latlim, type = "alt", intMethod = "lin",
  boxlon = 4, boxlat = 4, delta = 0)
}
\arguments{
\item{dat}{...}

\item{TG}{...}

\item{mask}{}

\item{lonlim}{...}

\item{latlim}{...}

\item{type}{...}

\item{intMethod}{...}

\item{boxlon}{...}

\item{boxlat}{...}

\item{delta}{...}
}
\value{
matrix with land values
##'
}
\description{
Extract MDT values at TG positions
}
\details{
...
}

