% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getSubGrid.R
\name{getSubGrid}
\alias{getSubGrid}
\title{Extract sub grid}
\usage{
getSubGrid(grid, lonlim, latlim, glonlim = c(0, 360), glatlim = c(-90, 90),
  res = 0.125)
}
\arguments{
\item{grid}{Input grid[longitude,latitude] of type matrix.}

\item{lonlim}{Vector of length two containing the longitude limits of the sub grid.}

\item{latlim}{Vector of length two containing the latitude limits of the sub grid.}

\item{glonlim}{Vector of length two containing the longitude limits of the input grid. The default is c(0,360).}

\item{glatlim}{Vector of length two containing the latitude limits of the input grid. The default is c(-90,90).}

\item{res}{The resolution of the input grit in decimal degrees.}
}
\value{
List with the elements;  matrix g (sub grid), vector lon (longitudes), vector lat (latitudes)
##'
}
\description{
Extract sub grid
}
\details{
...
}
\examples{
data(landmask8_anyland)
out<-getSubGrid(landmask8_anyland,c(280,300),c(30,60),glonlim=c(0,360-0.125),glatlim=c(-90,90-0.125))
image(out$lon,out$lat,out$g)
}

