% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getTGCoast.R, R/helper.R
\name{getTGCoast}
\alias{getTGCoast}
\title{Helper function: Identify Coast lines with tide gauges}
\usage{
getTGCoast(polyCoast, TG, lonlim, latlim)

getTGCoast(polyCoast, TG, lonlim, latlim)
}
\arguments{
\item{polyCoast}{...}

\item{TG}{...}

\item{lonlim}{...}

\item{latlim}{...}

\item{polyCoast}{...}

\item{TG}{...}

\item{lonlim}{...}

\item{latlim}{...}
}
\value{
list with mean and sd values
##'

coastid for each TG
}
\description{
Helper function: Identify Coast lines with tide gauges

Find index of coast line with Tide gauges
}
\details{
...
}

