% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helper.R
\name{getTGid}
\alias{getTGid}
\title{Helper function: Finds row and col id of Tide gauges}
\usage{
getTGid(TG, lonlim, latlim)
}
\arguments{
\item{TG}{Data frame or matrix with tide gauge information. The dimension of TG is N x 4, where N is the number of tide gauges and the 4 columns are; PSMSL station id, latitude, longitude, MDT value at the tide gauge.}

\item{lonlim}{Vector of length 2 with the longitude data grid limits, c(lonlim[1],lonlim[2]). The limits must be given in whole degrees.}

\item{latlim}{Vector of length 2 with the longitude data grid limits, c(lonlim[1],lonlim[2]). The limits must be given in whole degrees.}
}
\value{
Matrix[N,2] with row and coulmn id for the tide gauges
}
\description{
Helper function: Finds row and col id of Tide gauges
}

