% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid2file.R
\name{grid2file}
\alias{grid2file}
\title{Write a grid[lon,lat] to a netcdf file}
\usage{
grid2file(grid, varname = "MDT", filename = "grid.nc")
}
\arguments{
\item{grid}{An object as returned by the function 'iterativeAveSmoother', which includes a list containing  a matrix g[lon,lat], a vector lon (longitudes) and a vector lat (latitudes).}

\item{varname}{A string containing the name of the variable, the default is 'MDT'}

\item{filename}{A string containing the file name, the default is 'grid.nc'}
}
\description{
This function \code{grid2file} saves a grid[lon,lat] to a netcdf file.
}
\details{
...
}
\examples{
\dontrun{
grid2file(boxTG_ALT,filename='MDT_filtered.nc')
}
}
