% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/iterativeAveSmoother.R
\name{iterativeAveSmoother}
\alias{iterativeAveSmoother}
\title{Iterative box filter}
\usage{
iterativeAveSmoother(dat, mask, TGmat, latlim, radius = 0.15/0.83, nit = 10,
  dlat = 0.125)
}
\arguments{
\item{dat}{Matrix[lon,lat] to be filtered}

\item{mask}{Matrix[lon,lat] containing land/ocean values. Land has value 0}

\item{TGmat}{Matrix[lat,lon] containing tide gauge values}

\item{latlim}{Vector of length 2, c(lat1,lat2), containing latitude limits of the matrix dat}

\item{radius}{Filter radius}

\item{nit}{Number of iterations of the box filter}

\item{dlat}{Grid spacing of the matrix dat}
}
\value{
out[longitude,latitude]
##'
}
\description{
This function \code{iterativeAveSmoother} ........
}
\details{
...
}

