\name{MDT_EGM08_2003_2007}
\alias{MDT_EGM08_2003_2007}
\docType{data}
\title{MDT_EGM08_2003_2007}
\description{
  Estimated MDT values at selected tide gauges. These are based on the
  estimated mean sea level at the gauges and the EGM 2008 geoid model.
  The MDT estimates are in the .... tide system.

}
\usage{data(MDT_EGM08_2003_2007)}
\format{
 A data frame with 240 observations on the following 4 variables.
  \describe{
    \item{\code{PSMSLid}}{a numeric vector}
    \item{\code{Lat}}{a numeric vector}
    \item{\code{Lon}}{a numeric vector}
    \item{\code{MDT}}{a numeric vector}
  }   
 }
\details{
}
\source{
...
}
\references{
Create by Mederic Gravelle 
  
}
\examples{
data(MDT_EGM08_2003_2007)
}
\keyword{datasets}
