\name{MDT_eigen6c4r_2003_2007}
\alias{MDT_eigen6c4r_2003_2007}
\docType{data}
\title{MDT_eigen6c4r_2003_2007}
\description{
  Estimated MDT values at selected tide gauges. These are based on the
  estimated mean sea level at the gauges and the eigen6c4r geoid model.
  The MSS estimates are relative to the WGS84 ellipsoid.

}
\usage{data(MDT_eigen6c4r_2003_2007)}
\format{
 A data frame with 240 observations on the following 4 variables.
  \describe{
    \item{\code{PSMSLid}}{a numeric vector}
    \item{\code{Lat}}{a numeric vector}
    \item{\code{Lon}}{a numeric vector}
    \item{\code{MDT}}{a numeric vector}
  }   
 }
\details{
}
\source{
...
}
\references{
Create by Mederic Gravelle and Ole B. Andersen 
  
}
\examples{
data(MDT_eigen6c4r_2003_2007)
}
\keyword{datasets}
