\name{mean2TF_AddThis}
\alias{mean2TF}
\docType{data}
\title{
mean2TF_AddThis
}
\description{
Grid to go from mean tide to tide free. The grid should be added. mean2TF_AddThis is a 2880 x 1440 matrix.
}

\usage{data("mean2TF_AddThis")}
\format{
 A matrix[lon,lat] of dimension 2880 x 1440 
}
\details{
...
}
\source{
...
}
\references{
Created by Ole B. Andersen \email{oa@space.dtu.dk}
}
\examples{
data(mean2TF_AddThis)
}
\keyword{datasets}
