% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRegGridBin.R
\name{readRegGridBin}
\alias{readRegGridBin}
\title{Function that reads a regular global binary grid.}
\usage{
readRegGridBin(filename, nx = 2880, ny = 1440, res = 0.125, ...)
}
\arguments{
\item{filename}{A character string with the file name.}

\item{nx}{Number of rows in the grid (longitude). Deafault is nx=2880.}

\item{ny}{Number of columns in the grid (latitude). Deafault is ny=1440.}

\item{res}{The grid spacing in degrees. Default is res=0.125.}

\item{...}{Additional arguments to readBin.}
}
\value{
An object of the type coastMDT;  a list containing a matrix g[lon,lat], a vector lon (longitudes) and a vector lat (latitudes).
}
\description{
Function that reads a regular global binary grid.
}
\details{
The default is grid is a 1/8 degree grid which is 2880 longitudes by 1440 latitudes: longitudes (0.5,1.5,2.5 .... 2879.5)/8 degrees, latitudes  -90 + (0.5,1.5,2.5 .... 1439.5)/8 degrees. The type is real*4, hence the length of the file i 4*nx*ny. The unit is meter.
}
