\name{TF2mean_AddThis}
\alias{TF2mean}
\docType{data}
\title{
TF2mean_AddThis
}
\description{
to be written
}
\usage{data("TF2mean_AddThis")}
\format{
  A data frame with 8476 observations on the following 4 variables.
  \describe{
    \item{\code{height}}{a numeric vector}
    \item{\code{lake}}{a numeric vector}
    \item{\code{track}}{a numeric vector}
    \item{\code{time}}{a numeric vector}
  }
}
\details{
...
}
\source{
...
}
\references{
...
}
\examples{
data(TF2mean_AddThis)
}
\keyword{datasets}
