% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoidTF2MT.R
\name{tideConvert}
\alias{tideConvert}
\title{Function for converting between different permanent tide systems.}
\usage{
tideConvert(phi, convtype, k = 0.3)
}
\arguments{
\item{phi}{The latitude in degrees.}

\item{convtype}{A character string giving the type of conversion. The legal strings are; 'MT2ZT', 'ZT2MT', 'ZT2TF', 'TF2ZT', 'MT2TF', and 'TF2MT'. Here, MT is mean tide, ZT is zero tide and TF is tide free (or nontidal)}

\item{k}{is a love number, the default value is k=0.3}
}
\value{
The conversion correction; An array of height differences in meters.
}
\description{
Function for converting between different permanent tide systems.
}
\details{
The conversion expressions are based on Ekman, 1989. The correction must be added.
}
