% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wgs2topCorr.R
\name{wgs2topCorr}
\alias{wgs2topCorr}
\title{Function that estiamtes the height difference between WGS84 and Topex ellipsoids}
\usage{
wgs2topCorr(phi)
}
\arguments{
\item{phi}{The latitude in degrees.}
}
\value{
An array with the height differences.
}
\description{
Function that estiamtes the height difference between WGS84 and Topex ellipsoids
}
\details{
...
}

