% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compareWithTG.R
\name{compareWithTG}
\alias{compareWithTG}
\title{Extract MDT values at TG positions}
\usage{
compareWithTG(TG, dat, lonlim, latlim, boxlon = 3, boxlat = 3)
}
\arguments{
\item{TG}{Data frame or matrix with tide gauge information. The dimension of TG is N x 4, where N is the number of tide gauges and the 4 columns are; PSMSL station id, latitude, longitude, MDT value at the tide gauges.}

\item{dat}{Matrix[lon,lat] with MDT values}

\item{lonlim}{Vector of length 2 with the longitude data grid limits, c(lonlim[1],lonlim[2]). The limits must be given in whole degrees.}

\item{latlim}{Vector of length 2 with the longitude data grid limits, c(lonlim[1],lonlim[2]). The limits must be given in whole degrees.}

\item{boxlon}{The number ((2 x boxlon) +1) of grid cells in the longitude direction, that is used to estimate the altimetry based MDT value at the coast.}

\item{boxlat}{The number ((2 x boxlat) +1) of grid cells in the latitude direction, that is used to estimate the altimetry based MDT value at the coast.}
}
\value{
A list that includes:
mean: The mean values of the field in the box, defined by boxlon and boxlat at each tide gauge position.
sd: The standard deviation of the field in the box, defined by boxlon and boxlat at each tide gauge position.
bias: The bias between the mean field values and the tide gauge MDT values.
diff: The difference between the mean field values and the bias corrected tide gauge MDT values
RMS: The RMS of the mean field values and the tide gauge MDT values
}
\description{
The function extracts MDT values at the position of the tide gauges, and compares the MDT values of the field with MDT values based on the tide gauges.
}
\details{
Besides the list, that is returned. A plot of the difference between the altimetry and the tide gauges MDT values is automatically generated
}
\examples{
#load data
data(landmask8) #land mask
data(difmss15eig6c4r)
data(MDT_EGM08_2003_2007)
TG<-MDT_EGM08_2003_2007
#region of interest
lonlim<-c(275,300)
latlim<-c(20,55)
#sub grids
mask<-getSubGrid(landmask8,lonlim,latlim)
rawUS<-getSubGrid(difmss15eig6c4r,lonlim,latlim)
idraw<-which(mask$g==0,arr.ind=TRUE)
rawUS$g[idraw]<-NA
TGsub<-getSubTG(TG,lonlim,latlim)
#Compare with tide gauges
res<-compareWithTG(TGsub,rawUS$g,lonlim,latlim)
}

