% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compareWithTG.R
\name{compareWithTG}
\alias{compareWithTG}
\title{Extract MDT values at TG positions}
\usage{
compareWithTG(TG, dat, lonlim, latlim, boxlon = 3, boxlat = 3,
  export = FALSE, tgfile = "TGcompare.csv")
}
\arguments{
\item{TG}{Data frame or matrix with tide gauge information. TG should contain at least the columns with the names 'Longitude', 'Latitude', and 'TGMDT'. 'TGMDT' should contain MDT values at the tide gauge positions.}

\item{dat}{An object as returned by the function 'getSubGrid' or 'iterativeAveSmoother', which includes a list containing  a matrix g[lon,lat], a vector lon (longitudes) and a vector lat (latitudes).}

\item{lonlim}{Vector of length 2 with the longitude data grid limits, c(lonlim[1],lonlim[2]). The limits must be given in whole degrees.}

\item{latlim}{Vector of length 2 with the longitude data grid limits, c(lonlim[1],lonlim[2]). The limits must be given in whole degrees.}

\item{boxlon}{The number ((2 x boxlon) +1) of grid cells in the longitude direction, that is used to estimate the altimetry based MDT value at the coast.}

\item{boxlat}{The number ((2 x boxlat) +1) of grid cells in the latitude direction, that is used to estimate the altimetry based MDT value at the coast.}

\item{export}{If true the information estimated in the function is saved in a csv file. The default name is "TGcompare.csv". The file contains; All columns in the data frame TG,Alt_mean MDT,MDT_Alt_sd,bias corrected difference (alt-TG_bias_corr).}

\item{tgfile}{a character string giving the name of the file.}
}
\value{
A list that includes:
mean: The mean values of the field in the box, defined by boxlon and boxlat at each tide gauge position.
sd: The standard deviation of the field in the box, defined by boxlon and boxlat at each tide gauge position.
bias: The bias between the mean field values and the tide gauge MDT values.
diff: The difference between the mean field values and the bias corrected tide gauge MDT values
RMS: The RMS of the mean field values and the tide gauge MDT values
}
\description{
The function extracts MDT values at the position of the tide gauges, and compares the MDT values of the field with MDT values based on the tide gauges.
}
\details{
Besides the list, that is returned. A plot of the difference between the altimetry and the tide gauges MDT values is automatically generated
}

