% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helper.R
\name{getBoxMean}
\alias{getBoxMean}
\title{Helper function: Find mean value in a box}
\usage{
getBoxMean(dat, id, boxlon = 4, boxlat = 4)
}
\arguments{
\item{dat}{Matrix[lon,lat] with MDT values}

\item{id}{Matrix[N,2] with row and column id, reprecenting the center of the box}

\item{boxlon}{Integer. The number ((2 x boxlon) +1) of grid cells in the longitude direction, that is used to estimate the altimetry based MDT value at the coast.}

\item{boxlat}{Integer. The number ((2 x boxlat) +1) of grid cells in the latitude direction, that is used to estimate the altimetry based MDT value at the coast.}
}
\value{
list with mean and sd values
}
\description{
Helper function: Find mean value in a box
}
\details{
...
}

