% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getLandInfo.R
\name{getLandInfo}
\alias{getLandInfo}
\title{Estimate land values}
\usage{
getLandInfo(mycoast, mask, dat, delta = 0, boxlon = 4, boxlat = 4)
}
\arguments{
\item{mycoast}{...}

\item{mask}{...}

\item{dat}{...}

\item{delta}{...}

\item{boxlon}{}

\item{boxlat}{}

\item{scale}{}
}
\value{
Matrix myland[longitude,latitude]
##'
}
\description{
Estimate land values
}
\details{
...
}

