% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getSubGrid.R
\name{getSubGrid}
\alias{getSubGrid}
\title{Extract sub grid}
\usage{
getSubGrid(grid, lonlim, latlim, res = 0.125, glonlim = c(0 + (res/2), 360 -
  (res/2)), glatlim = c(-90 + (res/2), 90 - (res/2)))
}
\arguments{
\item{grid}{Input grid[longitude,latitude] of type matrix.}

\item{lonlim}{Vector of length two containing the longitude limits of the sub grid. The limits must be given with the smallest longitude first for example c(270,300), except when the 0 longitude id crossed. In this case for example c(355,10). Acceptable values are between 0 and 360.}

\item{latlim}{Vector of length two containing the latitude limits of the sub grid.}

\item{res}{The resolution of the input grit in decimal degrees.}

\item{glonlim}{Vector of length two containing the longitude limits of the input grid. The default is c(0,360).}

\item{glatlim}{Vector of length two containing the latitude limits of the input grid. The default is c(-90,90).}
}
\value{
List with the elements;  matrix g (sub grid), vector lon (longitudes), vector lat (latitudes)
##'
}
\description{
Extract sub grid
}
\details{
...
}
\examples{
data(landmask8)
out<-getSubGrid(landmask8,c(280,300),c(30,60))
image(out$lon,out$lat,out$g)
}

