% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helper.R
\name{getSubTG}
\alias{getSubTG}
\title{Find subset of Tide gauges}
\usage{
getSubTG(TG, lonlim, latlim)
}
\arguments{
\item{TG}{...}

\item{lonlim}{...}

\item{latlim}{...}
}
\value{
subset of TG
}
\description{
Find subset of Tide gauges
}

