% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{getSubTG}
\alias{getSubTG}
\title{Find subset of tide gauges}
\usage{
getSubTG(TG, lonlim, latlim)
}
\arguments{
\item{TG}{Data frame or matrix with tide gauge information. TG should contain at least the columns with the names 'Longitude' and 'Latitude'.}

\item{lonlim}{Vector of length 2 with the longitude data grid limits, c(lonlim[1],lonlim[2]). The limits must be given in whole degrees.}

\item{latlim}{Vector of length 2 with the longitude data grid limits, c(lonlim[1],lonlim[2]). The limits must be given in whole degrees.}
}
\value{
subset of TG
}
\description{
Find subset of tide gauges
}
