% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{getTGid}
\alias{getTGid}
\title{Helper function: Finds row and col id of Tide gauges}
\usage{
getTGid(TG, lonlim, latlim)
}
\arguments{
\item{TG}{Data frame or matrix with tide gauge information. TG should contain at least the columns with the names 'Longitude', 'Latitude', and 'TGMDT'. 'TGMDT' should contain MDT values at the tide gauge positions.}

\item{lonlim}{Vector of length 2 with the longitude data grid limits, c(lonlim[1],lonlim[2]). The limits must be given in whole degrees.}

\item{latlim}{Vector of length 2 with the longitude data grid limits, c(lonlim[1],lonlim[2]). The limits must be given in whole degrees.}
}
\value{
Matrix[N,2] with row and coulmn id for the tide gauges
}
\description{
Helper function: Finds row and col id of Tide gauges
}
