% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helper.R
\name{getTGid}
\alias{getTGid}
\title{Find row and col id of Tide gauges}
\usage{
getTGid(TG, lonlim, latlim)
}
\arguments{
\item{TG}{...}

\item{lonlim}{...}

\item{latlim}{...}
}
\value{
TGid
}
\description{
Find row and col id of Tide gauges
}

