% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getTGVal.R
\name{getTGVal}
\alias{getTGVal}
\title{Extract MDT values at TG positions}
\usage{
getTGVal(TG, dat, mask, lonlim, latlim, boxlon = 4, boxlat = 4)
}
\arguments{
\item{TG}{...}

\item{dat}{...}

\item{lonlim}{...}

\item{latlim}{...}

\item{boxlon}{}

\item{boxlat}{}
}
\value{
list with mean and sd values
##'
}
\description{
Extract MDT values at TG positions
}
\details{
...
}

