% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/iterativeAveSmoother.R
\name{iterativeAveSmoother}
\alias{iterativeAveSmoother}
\title{Iterative box filter}
\usage{
iterativeAveSmoother(dat, mask, land, latlim, radius = 0.15/0.83, nit = 10,
  res = 0.125)
}
\arguments{
\item{dat}{Matrix[lon,lat] to be filtered}

\item{mask}{Matrix[lon,lat] containing land/ocean values. Land=0, water=1.}

\item{land}{Matrix[lat,lon] containing land values}

\item{latlim}{Vector of length 2 with the longitude data grid limits, c(lonlim[1],lonlim[2]). The limits must be given in whole degrees.}

\item{radius}{Filter radius. Default is radius=0.15/0.83}

\item{nit}{Number of iterations of the box filter. Default is nit=10}

\item{dlat}{Grid spacing of the matrix dat. Default is dlat=0.125}
}
\value{
Matrix[lat,lon] with filtered MDT values.
}
\description{
This function \code{iterativeAveSmoother} ........
}
\details{
...
}

