% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/iterativeAveSmoother.R
\name{iterativeAveSmoother}
\alias{iterativeAveSmoother}
\title{Iterative box filter}
\usage{
iterativeAveSmoother(dat, mask, land, radius = 0.15/0.83, nit = 10,
  res = 0.125)
}
\arguments{
\item{dat}{An object as returned by the function 'getSubGrid', which includes a list containing  a matrix g[lon,lat], a vector lon (longitudes) and a vector lat (latitudes). The matrix dat$g[lon,lat] containes the values to be filtered.}

\item{mask}{An object as returned by the function 'getSubGrid', which includes a list containing  a matrix g[lon,lat], a vector lon (longitudes) and a vector lat (latitudes). mask$g is a Matrix[lon,lat] representing the land mask, where land=0 and water=1.}

\item{land}{Matrix[lon,lat] containing land values}

\item{radius}{Filter radius. Default is radius=0.15/0.83}

\item{nit}{Number of iterations of the box filter. Default is nit=10}

\item{dlat}{Grid spacing of the matrix dat. Default is dlat=0.125}
}
\value{
List with the elements;  matrix[lon,lat] g (grid), vector lon (longitudes), vector lat (latitudes).
}
\description{
The function \code{iterativeAveSmoother} is a simple average filter applied nit number of times. The size of the filter in the E-W direction is scaled according to the latitude.
}
\details{
...
}

