\name{mean2TF_AddThis}
\alias{mean2TF}
\docType{data}
\title{
mean2TF_AddThis
}
\description{
to be written
}
\usage{data("mean2TF_AddThis")}
\format{
  A data frame with 8476 observations on the following 4 variables.
  \describe{
    \item{\code{height}}{a numeric vector}
    \item{\code{lake}}{a numeric vector}
    \item{\code{track}}{a numeric vector}
    \item{\code{time}}{a numeric vector}
  }
}
\details{
...
}
\source{
...
}
\references{
...
}
\examples{
data(mean2TF_AddThis)
}
\keyword{datasets}
