% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotMDT.R
\name{plotMDT}
\alias{plotMDT}
\title{Plot MDT grid}
\usage{
plotMDT(mymdt, zlim, lonlim, latlim, addContour = TRUE, res = 0.125,
  conlev = 0.05, ...)
}
\arguments{
\item{mymdt}{...}

\item{zlim}{...}

\item{lonlim}{...}

\item{latlim}{...}

\item{addContour}{...}

\item{res}{...}

\item{conlev}{...}

\item{...}{...}
}
\description{
Plot MDT grid
}
\details{
...
}

