% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotMDT.R
\name{plotMDT}
\alias{plotMDT}
\title{Plot MDT grid}
\usage{
plotMDT(dat, zlim, addContour = TRUE, conlev = 0.05, ...)
}
\arguments{
\item{dat}{An object as returned by the function 'getSubGrid' or 'iterativeAveSmoother', which includes a list containing  a matrix g[lon,lat], a vector lon (longitudes) and a vector lat (latitudes).}

\item{zlim}{Range of the MDT values given as a  Vector of length 2.}

\item{addContour}{Bolean; To add a contour plot. Default is TRUE}

\item{conlev}{The spacing between contour lines given in meters. The default is 0.05.}

\item{...}{Additional arguments to image.plot from fields}
}
\description{
Plot MDT grid
}
\details{
...
}

