% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readncdf1var.R
\name{readncdf1var}
\alias{readncdf1var}
\title{Read net cdf file with one variable}
\usage{
readncdf1var(filename)
}
\arguments{
\item{filename}{String containing the filename}
}
\value{
Matrix var[longitude,latitude]
##' @details ...
}
\description{
This function \code{readncdf1var} ........
}
\examples{
\dontrun{
mydat<-readncdf1var('landmask8.nc')
}
}
